requirements.txt
python-i18n==0.3.9
pytz==2022.1